+++ RELEASE_NOTES.md
@@ -1,6 +1,7 @@
 # 3.10.0
 - New Features
   - Handle.getJdbi gets owning Jdbi instance
+  - sqlobject's `EmptyHandling` enum backported to core for invocations of `SqlStatement.bindList`
 - Bug Fixes
   - onDemand invocations @CreateSqlObject create new on-demand SqlObjects
   - onDemand SqlObject.withHandle / Transactional.inTransaction are now safe to call even outside an on-demand context